requirements.txt
fastapi
uvicorn
python-dotenv
pillow
pdf2image
opencv-python-headless
numpy
google-genai
python-multipart